backend/requirements.txt
Flask==2.0.1
Flask-SocketIO==5.1.1
Flask-CORS==4.0.0
python-dotenv==0.19.0
eventlet==0.33.0
Werkzeug==2.0.1
python-engineio==4.2.1
python-socketio==5.4.0
uuid==1.30
redis==4.6.0
gunicorn==21.2.0 